backend/api/permissions.py
from rest_framework.permissions import SAFE_METHODS, BasePermission


class IsAuthorOrReadOnly(BasePermission):
    """
    Access is allowed only to  administrators, authors.
    Other users can only view objects, but not edit or delete them.
    """

    def has_object_permission(self, request, view, obj):
        return request.method in SAFE_METHODS or request.user == obj.author
